requirements.txt
#Python dependencies
click
pyyaml